+++ esw-agent-service/esw-agent-service-api/shared/src/main/scala/esw/agent/service/api/protocol/AgentServiceRequest.scala
@@ -1,9 +1,9 @@
 package esw.agent.service.api.protocol
 
-import java.nio.file.Path
-
 import csw.location.api.models.ComponentId
 import csw.prefix.models.Prefix
+
+import java.nio.file.Path
 
 sealed trait AgentServiceRequest
 
@@ -14,6 +14,8 @@ object AgentServiceRequest {
 
   case class SpawnSequenceComponent(agentPrefix: Prefix, componentName: String, version: Option[String])
       extends AgentServiceRequest
+
+  case class SpawnContainers(agentPrefix: Prefix, hostConfigPath: String, isConfigLocal: Boolean) extends AgentServiceRequest
 
   case class KillComponent(componentId: ComponentId) extends AgentServiceRequest
 }