+++ esw-agent-service/esw-agent-service-app/src/main/scala/esw/agent/service/app/handlers/AgentServicePostHandler.scala
@@ -6,7 +6,12 @@ import csw.aas.http.SecurityDirectives
 import esw.agent.service.api.AgentServiceApi
 import esw.agent.service.api.codecs.AgentServiceCodecs._
 import esw.agent.service.api.protocol.AgentServiceRequest
-import esw.agent.service.api.protocol.AgentServiceRequest.{KillComponent, SpawnSequenceComponent, SpawnSequenceManager}
+import esw.agent.service.api.protocol.AgentServiceRequest.{
+  KillComponent,
+  SpawnContainers,
+  SpawnSequenceComponent,
+  SpawnSequenceManager
+}
 import esw.commons.auth.AuthPolicies
 import msocket.http.post.{HttpPostHandler, ServerHttpCodecs}
 
@@ -22,6 +27,9 @@ class AgentServicePostHandler(agentService: AgentServiceApi, securityDirective: 
 
       case SpawnSequenceManager(agentPrefix, obsModeConfigPath, isConfigLocal, version) =>
         sPost(complete(spawnSequenceManager(agentPrefix, obsModeConfigPath, isConfigLocal, version)))
+
+      case SpawnContainers(agentPrefix, hostConfigPath, isConfigLocal) =>
+        sPost(complete(spawnContainers(agentPrefix, hostConfigPath, isConfigLocal)))
 
       case KillComponent(componentId) =>
         sPost(complete(killComponent(componentId)))