+++ esw-agent-service/esw-agent-service-impl/src/main/scala/esw/agent/service/impl/AgentServiceImpl.scala
@@ -1,17 +1,16 @@
 package esw.agent.service.impl
 
-import java.nio.file.Path
-
 import akka.actor.typed.ActorSystem
 import csw.location.api.models.{ComponentId, Location}
 import csw.prefix.models.Prefix
 import esw.agent.akka.client.AgentClient
 import esw.agent.service.api.AgentServiceApi
-import esw.agent.service.api.models.{Failed, KillResponse, Killed, SpawnResponse}
+import esw.agent.service.api.models._
 import esw.commons.extensions.FutureEitherExt.FutureEitherOps
 import esw.commons.extensions.ListEitherExt.ListEitherOps
 import esw.commons.utils.location.LocationServiceUtil
 
+import java.nio.file.Path
 import scala.concurrent.{ExecutionContext, Future}
 
 class AgentServiceImpl(locationServiceUtil: LocationServiceUtil)(implicit actorSystem: ActorSystem[_]) extends AgentServiceApi {
@@ -36,6 +35,13 @@ class AgentServiceImpl(locationServiceUtil: LocationServiceUtil)(implicit actorS
     agentClient(agentPrefix)
       .flatMapRight(_.spawnSequenceComponent(componentName, version))
       .mapToAdt(identity, Failed)
+
+  override def spawnContainers(
+      agentPrefix: Prefix,
+      hostConfigPath: String,
+      isConfigLocal: Boolean
+  ): Future[SpawnContainersResponse] =
+    agentClient(agentPrefix).flatMapRight(_.spawnContainers(hostConfigPath, isConfigLocal)).mapToAdt(identity, Failed)
 
   override def killComponent(componentId: ComponentId): Future[KillResponse] = {
     locationServiceUtil